refactor: pull request code review
Change default file size for multiupload (10 -> 100 mb)

move ChunksUploader
