+++ lib/uploadcare/uploadcare_settings.rb
@@ -4,5 +4,5 @@ module Uploadcare
   PUBLIC_KEY = ENV.fetch('UPLOADCARE_PUBLIC_KEY')
   SECRET_KEY = ENV.fetch('UPLOADCARE_SECRET_KEY')
   AUTH_TYPE = 'Uploadcare.Simple'
-  MULTIPART_SIZE_THRESHOLD = 10 * 1024 * 1024
+  MULTIPART_SIZE_THRESHOLD = 100 * 1024 * 1024
 end
